rename the directory with repositories
